% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_GE.R
\name{adjust_GE}
\alias{adjust_GE}
\title{Function to select only the probes with highest expression for each gene and
remove the rest from the expression matrix}
\usage{
adjust_GE(eset)
}
\arguments{
\item{eset}{ExpressionSet object that has expression matrix in AssayData()}

\item{gene_symbols}{Character vector that matches rownames of exprs(eset)}
}
\description{
Function to select only the probes with highest expression for each gene and
remove the rest from the expression matrix
}
