% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hai.R
\name{adjust_hai}
\alias{adjust_hai}
\title{Function to generate HAI data table from ImmuneSpace Connection}
\usage{
adjust_hai(sdy, rawdata)
}
\arguments{
\item{sdy}{ImmuneSpace Study Name}

\item{rawdata}{dataframe of raw HAI data}
}
\description{
Function to generate HAI data table from ImmuneSpace Connection
}
