% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Gene_Analysis.R
\name{arithMean_comb_validate}
\alias{arithMean_comb_validate}
\title{Generates Single Gene Analysis Scores}
\usage{
arithMean_comb_validate(GEM, genes_p = NULL, genes_n = NULL, scale = 1)
}
\arguments{
\item{GEM}{expression matrix}

\item{genes_p}{up-regulated genes}

\item{genes_n}{down-regulated genes}

\item{scale}{unknown}
}
\description{
Generates Single Gene Analysis Scores
}
