% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_v2_EH_f.R
\name{capwords}
\alias{capwords}
\title{---
title: "HIPC Datasets - Processing"
author: "Renaud Gaujoux & Stefan Avey"
date: "06/11/2014"
output:
 pdf_document:
   toc: true
   toc_depth: 1}
\usage{
capwords(s, strict = FALSE)
}
\description{
---
}

