% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GE.R
\name{getGE}
\alias{getGE}
\title{Function to return the prepackaged expression matrix held in library. Kept
outside of Rmd to allow for original code to be shown above.}
\usage{
getGE(sdy)
}
\arguments{
\item{sdy}{name of study}
}
\description{
Function to return the prepackaged expression matrix held in library. Kept
outside of Rmd to allow for original code to be shown above.
}
