% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSDY_f.R
\name{getSDY}
\alias{getSDY}
\title{Function loads hipc eset objects from rds files and performs GE adjustment and subsetting, if directed via params.}
\usage{
getSDY(rds_data_dir, sdy, group, response, adjusted = FALSE,
  baselineOnly = FALSE)
}
\arguments{
\item{rds_data_dir}{path to the directory containing the dataset .rds files}

\item{sdy}{ImmuneSpace study id}

\item{group}{sample subset to load, either young or old}

\item{response}{HAI discrete variable to load into variable "Response".
Responders are the samples in the last class, while non-responders are
the samples in the first class. Samples from the intermediate class (if any)
get an NA value.}

\item{adjusted}{if TRUE, baseline expression data is adjusted for age (but NOT gender) with a linear regression model.  (intercept + residual) are given as the adjusted values for the baseline data only and any other time points will have missing values (NA).  Default is TRUE for Old subjects but FALSE for young or all subjects. NOTE: Behavior is different than in previous version!}

\item{baselineOnly}{When `TRUE`, only baseline (day 0 or pre-vaccination) time points will be returned. Default is `FALSE` for compatibility with previous versions.}
}
\value{
an ExpressionSet object with only one matrix in assayData. Which matrix depends on the combination of the group (age) and whether adjusted is set to TRUE or FALSE.
}
\description{
Function loads hipc eset objects from rds files and performs GE adjustment and subsetting, if directed via params.
}

