% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sample_names.R
\name{hashmap}
\alias{hashmap}
\title{Function to generate HAI data table from ImmuneSpace Connection}
\usage{
hashmap(keys, values, vec_2_hash)
}
\arguments{
\item{keys}{keys}

\item{values}{Values}

\item{vec_2_hash}{Vector to be hashed}
}
\description{
Function to generate HAI data table from ImmuneSpace Connection
}

