% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_HIPC_analysis_f.R
\name{hipc_full_pipeline}
\alias{hipc_full_pipeline}
\title{Runs full pipeline from start to finish with user input required at certain points}
\usage{
hipc_full_pipeline()
}
\value{
text and pdf files wtih significant gene pathways
}
\description{
Runs full pipeline from start to finish with user input required at certain points
}

