% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_HIPC_analysis_f.R
\name{hipc_make_rds}
\alias{hipc_make_rds}
\title{Generates rds objects from studies given the study names and directories holding the rawdata.}
\usage{
hipc_make_rds(studies, hai_dir, ge_dir, rds_dir)
}
\arguments{
\item{studies}{a vector of characters holding the study names for which to generate eSets}

\item{hai_dir}{a path for the preprocessed HAI data}

\item{ge_dir}{a path for the preprocessed GE and demographic data}

\item{rds_dir}{a path for the output files (rds)}
}
\description{
Generates rds objects from studies given the study names and directories holding the rawdata.
}

