% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_HIPC_analysis_f.R
\name{hipc_meta_analysis}
\alias{hipc_meta_analysis}
\title{Perform HIPC Meta Analysis using given parameters}
\usage{
hipc_meta_analysis(rds_dir, cohort, orig_params = T, output_dir)
}
\arguments{
\item{rds_dir}{the path for the directory holding the RDS files with esets of each study}

\item{cohort}{the age cohort, either 'young' or 'old', to use in analysis}

\item{orig_params}{a boolean that defaults to TRUE, but when set to false allows the user to specify non-original parameters for the analysis}

\item{output_dir}{the path for the pdf and text outputs of the analysis}
}
\description{
Takes in various parameters related to the cutoff points for gene pathway selection, HAI endpoint selection,
adjustments to gene expression data, and output types.
}

