% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_HIPC_analysis_f.R
\name{hipc_preprocess}
\alias{hipc_preprocess}
\title{Generates TSV text tables for use in the hipc pipeline from ImmuneSpace data}
\usage{
hipc_preprocess(studies, hai_dir, ge_dir, rawdata_dir, orig_params = T)
}
\arguments{
\item{studies}{a vector of characters holding the study names for which to generate eSets}

\item{hai_dir}{a path for the preprocessed HAI data}

\item{ge_dir}{a path for the preprocessed GE and demographic data}

\item{rawdata_dir}{a path for directory to download rawdata into}
}
\description{
Generates TSV text tables for use in the hipc pipeline from ImmuneSpace data
}

