% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_v2_EH_f.R
\name{make_rds}
\alias{make_rds}
\title{Function to generate expressionSet object and save as Rds file}
\usage{
make_rds(sdy, ge_dir, combined_hai, output_dir)
}
\arguments{
\item{sdy}{Study}

\item{ge_dir}{Directory holding gene expression data tables}

\item{combined_hai}{Combined hai data frame}

\item{output_dir}{Output directory}
}
\description{
Function to generate expressionSet object and save as Rds file
}

