% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_MakeDemo_f.R
\name{makeDemo}
\alias{makeDemo}
\title{Function to generate demographic data table from ImmuneSpace connection}
\usage{
makeDemo(sdy, output_dir)
}
\arguments{
\item{sdy}{Study}

\item{output_dir}{output directory}
}
\description{
Function to generate demographic data table from ImmuneSpace connection
}

