% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_MakeGE_f.R
\name{makeGE}
\alias{makeGE}
\title{Function to generate Gene Expression Table from ImmuneSpace Connection}
\usage{
makeGE(sdy, yale.anno = "original", sdy80.anno = "original",
  sdy80.norm = F, output_dir, rawdata_dir)
}
\arguments{
\item{sdy}{study}

\item{yale.anno}{Yale Studies gene symbol annotation original, library, or manifest}

\item{sdy80.anno}{SDY80 gene symbol annotation original or library}

\item{sdy80.norm}{normalize SDY80 expression values according to pipeline, default FALSE}

\item{output_dir}{output director}

\item{rawdata_dir}{directory for holding rawdata downloads}
}
\description{
Function to generate Gene Expression Table from ImmuneSpace Connection
}

