% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Gene_Analysis.R
\name{makeGEM}
\alias{makeGEM}
\title{Converts expressionSet into list for MetaIntegrator analysis}
\usage{
makeGEM(eset, rmMods, ens_table)
}
\arguments{
\item{eset}{expressionSet object}

\item{rmMods}{Remove moderates, takes boolean}

\item{ens_table}{ensemble table of same name preloaded with pkg}
}
\description{
Converts expressionSet into list for MetaIntegrator analysis
}
