% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_MakeHAI_f.R
\name{makeHAI}
\alias{makeHAI}
\title{Function to generate HAI data table from ImmuneSpace Connection}
\usage{
makeHAI(sdy, output_dir)
}
\arguments{
\item{sdy}{Study}

\item{output_dir}{output directory}
}
\description{
Function to generate HAI data table from ImmuneSpace Connection
}

