% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_Analysis_Demo.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Function to perform meta analysis for HIPC ImmuneSignatures Project}
\usage{
meta_analysis(adj_eset_list, cohort, gene_set)
}
\arguments{
\item{cohort}{Study cohort, young or old}

\item{eset_list}{list of expressionSet objects holding GE, HAI data}
}
\description{
Function to perform meta analysis for HIPC ImmuneSignatures Project
}
