% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Gene_Analysis.R
\name{singleGeneAnalysis}
\alias{singleGeneAnalysis}
\title{Function to perform single gene meta analysis for HIPC ImmuneSignatures Project}
\usage{
singleGeneAnalysis(adjEsetList, cohort)
}
\arguments{
\item{adjEsetList}{list of expressionSet objects holding expr, pheno data adjusted for cohort}

\item{cohort}{Study cohort, young or old}
}
\description{
Function to perform single gene meta analysis for HIPC ImmuneSignatures Project
}
