% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Gene_Analysis.R
\name{swapCols}
\alias{swapCols}
\title{helper function SDY404 column swap}
\usage{
swapCols(em, sub1, sub2)
}
\arguments{
\item{em}{expression matrix with colnames including sub1 and sub2}

\item{sub1}{subject 1}

\item{sub2}{subject 2}
}
\description{
helper function SDY404 column swap
}
