% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_v2_EH_f.R
\name{combine_hai_data}
\alias{combine_hai_data}
\title{Function creates combined hai table and rds file from all six studies}
\usage{
combine_hai_data(hai_dir, output_dir)
}
\arguments{
\item{hai_dir}{Directory holding HAI data tables}

\item{output_dir}{Output Directory}
}
\description{
Function creates combined hai table and rds file from all six studies
}

