% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_Analysis_Demo.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Function to perform meta analysis for HIPC ImmuneSignatures Project}
\usage{
meta_analysis(geneSetDB, eset_list, cohort, FDR.cutoff = 0.5,
  pvalue.cutoff = 0.01, endPoint = "fc_res_max_d30")
}
\arguments{
\item{geneSetDB}{table defining gene sets}

\item{cohort}{Study cohort, young or old}

\item{FDR.cutoff}{Cutoff for q-values in selecting significant gene sets}

\item{pvalue.cutoff}{cutoff for p-values in selecting significant gene sets}

\item{endPoint}{HAI table column used for categorizing response}
}
\description{
Function to perform meta analysis for HIPC ImmuneSignatures Project
}

